Refs #21126 First pass at the system test
