+++ Testing/Data/SystemTest/BASIS/ELWIN/BASIS_63652_sqw.nxs.md5
@@ -1,0 +1,1 @@
+5233fbb6e769a810a69e74b41306fef3
