+++ Testing/SystemTests/tests/analysis/BASISTests.py
@@ -1,0 +1,122 @@
+from __future__ import (absolute_import, division, print_function)
+
+import stresstesting
+from mantid import config
+from mantid.simpleapi import (BASISDiffraction, Load, GroupWorkspaces, ElasticWindowMultiple)
+
+
+class PreppingMixin(object):
+    r"""Common code for tests classes
+    
+    """
+    def setUp(self):
+        self.config = {p: config[p] for p in ('default.facility',
+                                              'default.instrument',
+                                              'datasearch.directories')}
+        config['default.facility'] = 'SNS'
+        config['default.instrument'] = 'BASIS'
+        config.appendDataSearchSubDir('BASIS/'+self._subdir)
+
+    def tearDown(self):
+        config.update(self.config)
+
+
+class CrystalDiffractionTest(stresstesting.MantidStressTest, PreppingMixin):
+    """
+    Run a reduction for a scan of runs probing different orientations
+    of a crystal.
+    """
+
+    def __init__(self):
+        super(CrystalDiffractionTest, self).__init__()
+        self.config = None
+        self._subdir = 'BASISDiffraction'
+        self.setUp()
+
+    def requiredFiles(self):
+        return ['BASIS_Mask_default_diff.xml',
+                'BSS_74799_event.nxs',
+                'BSS_74800_event.nxs',
+                'BSS_64642_event.nxs',
+                'BSS_75527_event.nxs',
+                'BASISOrientedSample.nxs']
+
+    def runTest(self):
+        """
+        Override parent method, does the work of running the test
+        """
+        try:
+            BASISDiffraction(SingleCrystalDiffraction=True,
+                             RunNumbers='74799-74800',
+                             MaskFile='BASIS_Mask_default_diff.xml',
+                             VanadiumRuns='64642',
+                             BackgroundRun='75527',
+                             PsiAngleLog='SE50Rot',
+                             PsiOffset=-27.0,
+                             LatticeSizes=[10.71, 10.71, 10.71],
+                             LatticeAngles=[90.0, 90.0, 90.0],
+                             VectorU=[1, 1, 0],
+                             VectorV=[0, 0, 1],
+                             Uproj=[1, 1, 0],
+                             Vproj=[0, 0, 1],
+                             Wproj=[1, -1, 0],
+                             Nbins=300,
+                             OutputWorkspace='peaky')
+        finally:
+            self.tearDown()
+
+    def validate(self):
+        """
+        Inform of workspace output after runTest(), and associated file to
+        compare to.
+        :return: strings for workspace and file name
+        """
+        self.tolerance = 0.1
+        return 'peaky', 'BASISOrientedSample.nxs'
+
+
+class ElwinTest(stresstesting.MantidStressTest, PreppingMixin):
+    r"""ELWIN tab of the Indirect Inelastic Interface 
+    """
+    def __init__(self):
+        super(ElwinTest, self).__init__()
+        self.config = None
+        self._subdir = 'ELWIN'
+        self.setUp()
+
+    def requiredFiles(self):
+        return ['BSS_63652_event.nxs',
+                'BSS_63700_event.nxs',
+                'BASIS_elwin_eq2.nxs']
+
+    def runTest(self):
+        """
+        Override parent method, does the work of running the test
+        """
+        # Load files and create workspace group
+        try:
+            names = ('BSS_63652_event', 'BSS_63700_event')
+            [Load(Filename=name + '.nxs', OutputWorkspace=name) for name in names]
+            GroupWorkspaces(InputWorkspaces=names, OutputWorkspace='elwin_input')
+            ElasticWindowMultiple(InputWorkspaces='elwin_input',
+                                  IntegrationRangeStart=-0.0035,
+                                  IntegrationRangeEnd=0.0035,
+                                  BackgroundRangeStart=-0.1,
+                                  BackgroundRangeEnd=-0.05,
+                                  SampleEnvironmentLogName='SensorA',
+                                  SampleEnvironmentLogValue='average',
+                                  OutputInQ='outQ',
+                                  OutputInQSquared='outQ2',
+                                  OutputELF='ELF',
+                                  OutputELT='ELT')
+        finally:
+            self.tearDown()
+
+    def validate(self):
+        """
+        Inform of workspace output after runTest(), and associated file to
+        compare to.
+        :return: strings for workspace and file name
+        """
+        self.tolerance = 0.1
+        return 'outQ2', 'BASIS_elwin_eq2.nxs'
